Install app packages

Install packages for an app

Usage

twinkle_install_packages(name)

Arguments

name

Name of the app

Value

Invisibly, a list containing the sha that the installation was based on and the conan installation id from the installation