Sync app

Sync an app and its library to the staging or production shiny server.

Usage

twinkle_sync(name, production, verbose = TRUE)

Arguments

name

Name of the app

production

Logical, indicating if we want to update the production instance. If FALSE, then staging is updated.

verbose

Logical, indicating if output should be shown (passed to system commands, which are otherwise hard to capture output from)

Value

A named character vector with the ids of the repo (last SHA) and the library (conan id)